refactor: GoalDetails from TableViewController to UIViewController with UITableView and some minor updates in Agenda and AddingGoal
