+++ Agenda/View/GoalTableView.swift
@@ -1,0 +1,12 @@
+//
+//  GoalTableView.swift
+//  Agenda
+//
+//  Created by Егор Бадмаев on 03.01.2022.
+//
+
+import UIKit
+
+class GoalTableView: UIView {
+    
+}
